Update throttling example
